Refactors out highlight plugin
